+++ tuxemon/states/difficulty_pick.py
@@ -2,6 +2,7 @@
 # Copyright (c) 2014-2026 William Edwards <shadowapex@gmail.com>, Benjamin Bean <superman2k5@gmail.com>
 from __future__ import annotations
 
+from collections.abc import Callable
 from functools import partial
 from typing import ClassVar
 
@@ -10,32 +11,30 @@ from pygame_menu.widgets.selection.highlight import HighlightSelection
 
 from tuxemon.locale import T
 from tuxemon.menu.menu import PygameMenuState
-from tuxemon.platform.const.sizes import PLAYER_NPC
-from tuxemon.player import Player
 from tuxemon.prepare import SCREEN_SIZE
-from tuxemon.session import local_session
 
 DIFFICULTIES = ["beginner", "easy", "normal", "hard", "expert"]
 
 
-class DifficultyBattleState(PygameMenuState):
-    """
-    A state that allows players to choose the difficulty level before entering the battle.
-    """
+class DifficultyPickState(PygameMenuState):
+    """Generic difficulty selection state."""
 
-    name: ClassVar[str] = "DifficultyBattleState"
+    name: ClassVar[str] = "DifficultyPickState"
 
-    def __init__(self) -> None:
+    def __init__(
+        self,
+        on_pick: Callable[[str], None],
+        difficulties: list[str] = DIFFICULTIES,
+    ) -> None:
         width, height = SCREEN_SIZE
         super().__init__(height=height, width=width)
 
+        self.on_pick = on_pick
+        self.difficulties = difficulties
         self._build_menu()
         self.reset_theme()
 
     def _build_menu(self) -> None:
-        """
-        Constructs the difficulty selection menu with a title label and difficulty buttons.
-        """
         title = T.translate("choose_difficulty")
         self.menu.add.label(
             title=title,
@@ -44,22 +43,15 @@ class DifficultyBattleState(PygameMenuState):
             underline=True,
         )
 
-        for level in DIFFICULTIES:
+        for level in self.difficulties:
             self.menu.add.button(
                 title=T.translate(f"level_{level}"),
-                action=partial(self.start_battle, level),
+                action=partial(self._handle_pick, level),
                 button_id=f"diff_{level}",
                 font_size=self.font_type.medium,
                 selection_effect=HighlightSelection(),
                 align=locals.ALIGN_CENTER,
             )
 
-    def start_battle(self, difficulty: str) -> None:
-        Player.create(local_session, slug=PLAYER_NPC)
-        self.client.push_state(
-            "WorldState", session=local_session, map_name=None
-        )
-        self.client.event_engine.execute_action(
-            "set_variable", [f"difficulty:{difficulty}"]
-        )
-        self.client.event_engine.execute_action("load_yaml", ["battle_menu"])
+    def _handle_pick(self, difficulty: str) -> None:
+        self.on_pick(difficulty)
